Paths: skip edges for which no 'numeraire'-rate exists
When building DepthGraph, skip edges for which no 'numeraire'
exchange rate exists (rather than crashing). This can happen
if a symbol exists that is not connected to 'numeraire' (and
the RateMap thus doesn't contain a 'numeraire' exchange rate
for the given symbol).

Also, for the time being, use "trace" to log about when this
happens.